requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.1
django-filter==2.3.0
django-heroku==0.3.1
django-mptt==0.11.0
django-import-export==2.4.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
sqlparse==0.3.1
whitenoise==5.2.0
